Web-Service/Backend/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
asyncio
websockets
pinecone-client
python-dotenv
# google-generativeai==0.3.1
openai
transformers
sentence-transformers
langchain
langchain-community
langchain_pinecone
langchain_openai